stats module update

+++ statistics/stats.py
@@ -7,9 +7,9 @@ class Stats():
 
 	def median(x):
 		if len(x) % 2 != 0:
-			return sorted(list(x))[(len(x)+1)/2]
+			return sorted(list(x))[int((len(x)+1)/2)]
 		else:
-			return sum(sorted(list(x))[(len(x)/2):(len(x)/2)+2])/2
+			return sum(sorted(list(x))[int((len(x)/2)):int((len(x)/2)+2)])/2
 
 	def mode(x):
 		mode_val = 0; mode_count = 0;
